Renderer::renderPriorityとCamera2D::priorityの型の変更

+++ src/GE/Core/Camera2D.h
@@ -1,8 +1,9 @@
 ﻿#pragma once
 
+#include <vector>
+#include <cstdint>
 #include "Component.h"
 #include "GE/Render/RenderLayer.h"
-#include <vector>
 #include "Event.h"
 
 // 備考：
@@ -31,8 +32,8 @@ namespace GE
 		Camera2D(GameObject& gameObject);
 		~Camera2D() = default;
 
-		int GetPriority() const { return priority; }
-		void SetPriority(int priority) { this->priority = priority; }
+		int16_t GetPriority() const { return priority; }
+		void SetPriority(int16_t priority) { this->priority = priority; }
 
 		Render::RenderLayer::Bitmask GetCullingMask() const { return cullingMask; }
 		void SetCullingMask(Render::RenderLayer::Bitmask cullingMask) { this->cullingMask = cullingMask; }
@@ -46,7 +47,7 @@ namespace GE
 		/// <para>大きければ大きいほど画面の上に描画する。負数も許す。</para>
 		/// GE::Renderer::RendererのrenderPriorityと違う。
 		/// </summary>
-		int priority;
+		int16_t priority;
 		Render::RenderLayer::Bitmask cullingMask;
 
 	private:
