+++ src/GE/Render/Renderer.cpp
@@ -6,7 +6,7 @@ namespace GE::Render
 	Event<const Renderer&> Renderer::onCreated;
 	Event<const Renderer&> Renderer::onDestroying;
 
-	Renderer::Renderer(GameObject& gameObject, int renderPriority, RenderLayer::Bitmask renderLayer) :
+	Renderer::Renderer(GameObject& gameObject, int16_t renderPriority, RenderLayer::Bitmask renderLayer) :
 		Component(gameObject),
 		renderPriority(renderPriority),
 		renderLayer(renderLayer)
